Refactor dropdown button component
